requirements/base.txt
fastapi==0.70.0
mangum==0.12.3
uvicorn==0.15.0
PyJWT==2.3.0
